implemented definition of config search space
